add recordIntegration to pageview
