+++ public/index.html
@@ -12,7 +12,7 @@
     <script src="http://localhost:3000/index.js?appId=universal-tag-staging-test&test=true"></script>
     <script src="http://localhost:3000/index.js?appId=universal-tag-staging-test&environment=jane&retailId=segregate&test=true"></script> -->
     <!-- <script src="http://localhost:3000/index.js?appId=universal-tag-staging&event=signup&test=true&emailAddress=test123@gmail.com&hashedEmailAddress=test123@gmail.com"></script> -->
-    <!-- <script src="http://localhost:3000/index.js?appId=universal-tag-staging&environment=buddi&test=true"></script> -->
+    <script src="http://localhost:3000/index.js?appId=universal-tag-staging&environment=buddi&test=true&version=2.0"></script>
     <!-- Tracker END -->
 
     <!-- AWS Staging Tracker START -->
