+++ src/v2/index.ts
@@ -1,12 +1,14 @@
-import createTracker from "./snowplow/create-tracker";
+import createTracker from "./snowplow/events/create-tracker";
 import { QueryStringContext } from "../shared/types";
 import { debuggerPlugin } from "./snowplow/plugins";
 import { liquidMRetargetingPixel } from "./partners/liquidm/retargeting-pixel";
 import { tapadCookieSyncPixel } from "./partners/tapad/cookie-sync-pixel";
 import { tapadHashSyncPixel } from "./partners/tapad/hash-sync-pixel";
+import recordIntegration from "./snowplow/events/record";
 
 const applyV2 = (context: QueryStringContext): void => {
   createTracker(context);
+  recordIntegration(context);
   liquidMRetargetingPixel();
   tapadCookieSyncPixel();
   tapadHashSyncPixel();
